Adding tests, moving StartProcessAction and Executor to core libs
